Modify and extend permission handling on subshops and slots
Subshops have ability to auto close the shop when player attempts
to access them without permission via the "close-on-permission-fail"
option. This allows error message in console to be seen.

Items now have the ability to have a custom permission name specified
via the option "permission-name" which overrides the built in
"tokenmanager.use.shop.<SHOPNAME>-<SLOT #>" formatting. This allows
items to be tied to other permissions within the game and the same
permission to be used on multiple items.

Items now have the option to negate the permission check via the
new option "negative-permission". This reverses the permission check
so that items can only be purchased is a permission is not set. This
allows abilities and ingame items that are controlled via permissions
to be purchased and prevents repurchase without having to have a
second set of permissions for the shop, this is most useful when
combined with the "permission-name" option.